// apps/worker/src/location/generatedSubwayStations.ts
export interface SubwayStation {
  id: string;
  name: string;
  borough: string;
  latitude: number;
  longitude: number;
  lines: string[];
  gtfs_stop_ids: string;
}